requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
Django==1.11
django-cors-headers==2.0.2
django-filter==1.0.2
djangorestframework==3.4.6
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.10.0
djoser==0.5.4
drf-tracking==1.2.0
enum34==1.1.6
html5lib==0.999999999
idna==2.5
inflection==0.3.1
ipaddress==1.0.18
kiteconnect==3.6.0
Markdown==2.6.8
mechanize==0.3.5
more-itertools==3.0.0
ndg-httpsclient==0.4.2
numpy==1.12.1
packaging==16.8
pandas==0.20.0
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.5.0
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-utils==2.1.0
pytz==2017.2
Quandl==3.1.0
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.10
tqdm==4.14.0
webencodings==0.5.1
websocket-client==0.44.0
wget==3.2
whitenoise==3.3.0
